[IMP]Wireshark dependency update to 4.0; add out-of-bounds control to specify frame dissect result func;

+++ include/wireshark/ui/qt/models/tree_model_helpers.h
@@ -1,4 +1,4 @@
-/* tree_model_helpers.h
+/** @file
  *
  * Utility template classes for basic tree model functionality
  *
@@ -66,14 +66,14 @@ public:
 
     int childCount() const
     {
-        return childItems_.count();
+        return static_cast<int>(childItems_.count());
     }
 
     int row()
     {
         if (parent_)
         {
-            return parent_->childItems_.indexOf(VariantPointer<Item>::asQVariant((Item *)this));
+            return static_cast<int>(parent_->childItems_.indexOf(VariantPointer<Item>::asQVariant((Item *)this)));
         }
 
         return 0;
